skills page now showing skills
