+++ app/javascript/components/routes/Registration.js
@@ -2,44 +2,28 @@ import React from 'react'
 import Form from 'react-bootstrap/Form'
 import { useNavigate } from 'react-router-dom'
 import toast, { Toaster } from 'react-hot-toast'
-import userAPI from '../utils/api/user'
 import SkillTrackerButton from '../elements/button'
+import { useAppContext } from '../AppContext'
 
 const Registration = () => {
   const navigate = useNavigate()
 
-  const { userRegisterSubmit } = userAPI()
+  const { handleRegistrationSubmit } = useAppContext()
 
-  const handleSubmit = async (event) => {
-    event.preventDefault()
-
-    const formData = new FormData(event.target)
-
-    // Create an object to store the form data
-    const data = {}
-
-    for (const [name, value] of formData.entries()) {
-      data[name] = value
-    }
-
-    // Include the CSRF token in the form data
-    data.authenticity_token = document.querySelector('meta[name="csrf-token"]').content
-
+  const handleRegistration = (e) => {
     try {
-      const response = await userRegisterSubmit(data)
-      if (response) {
-        toast('Successfully registered')
-        navigate('/api/v1/login')
-      }
-    } catch (response) {
-      toast(response.error)
+      handleRegistrationSubmit(e)
+      toast('Successfully registered')
+      navigate('/api/v1/login')
+    } catch (error) {
+      toast(error)
     }
   }
 
   return (
     <div className="text-center d-flex-inline main-div">
       <h2 className="main-headers">Sign up</h2>
-      <Form onSubmit={handleSubmit} className="main-headers">
+      <Form onSubmit={(e) => { handleRegistration(e) }} className="main-headers">
         <Form.Group controlId="formFirstname">
           <Form.Label>First Name</Form.Label>
           <Form.Control type="name" name="first_name" className="text-center" autoFocus required />
