+++ app/javascript/components/routes/Skills.js
@@ -4,12 +4,14 @@ import PropTypes from 'prop-types'
 
 const Skills = () => {
   const { state } = useAppContext()
+  console.log('🚀 ~ file: Skills.js:7 ~ Skills ~ state:', state.skills)
 
   return (
     <>
       <div className="text-center">Skills</div>
       <ul>
         {
+          Object.keys(state.skills).length > 0 &&
           state.skills.map((skill) => {
             return (
               <li key={skill.id}>{skill.name}</li>
